+++ src/main/java/Seoul_Milk/sm_server/login/dto/request/JoinDTO.java
@@ -1,10 +1,12 @@
-package Seoul_Milk.sm_server.login.dto;
+package Seoul_Milk.sm_server.login.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
-public class LoginRequestDTO {
+@Setter
+public class JoinDTO {
     @Schema(description = "사번", example = "1111")
     private String employeeId;
     @Schema(description = "비밀번호", example = "1234")
