+++ src/main/java/Seoul_Milk/sm_server/login/dto/request/UpdatePwDTO.java
@@ -1,4 +1,4 @@
-package Seoul_Milk.sm_server.login.dto;
+package Seoul_Milk.sm_server.login.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
